Config file for using SuperPINDA together with Einsy Rambo + drop bad_region entries and tweak start and end position to avoid magnets through the power of math (#16)
* Add probe config for SuperPINDA on Einsy Rambo, without using faulty regions

After having several issues getting reliable bed meshes and first layers I moved the SuperPINDA back to the original pins on the Einsy Rambo, which seems to give more reliable results. Also tweaked the bed mesh geometry in order to completely miss the Mk52 magnet locations without using faulty regions.

This config file is only useful if you're using the SuperPINDA together with the "stock" SB toolend.

* Adjust max Y direction to avoid the final magnet

This tweak allows the SuperPINDA to avoid all the magnets on the plate.

* Fix incorrectly inverted SuperPINDA pin

* Revert "Fix incorrectly inverted SuperPINDA pin"

Moved this to a separate branch

This reverts commit ef39e83ab6cbfa4cf9b83b613a80db26869a3af5.

* Added comment explaining the reasoning behind the chosen mesh_min and mesh_max values